backend/requirements.txt
boto3
bcrypt
pyjwt
